// front_end/frontend_serv_vaccine/src/infrastructure/interfaces/auth-vac.response.ts
 // Generated by https://quicktype.io

export interface LoginResponse {
    statusCode: number;
    token:      string;
    usuario:    Usuario;
    more:       More;
    resp:       boolean;
    message:    string;
}

export interface More {
    _id:        ID;
    name:       string;
    lastname:   string;
    email:      string;
    birth:      string;
    gender_id:  string;
    status:     boolean;
    token:      string;
    phone:      string;
    isUser:     boolean;
    user_id:    ID;
    isChildren: boolean;
    age:        number;
    days_birth: number;
}

export interface ID {
    $oid: string;
}

export interface Usuario {
    _id:       ID;
    phone:     string;
    last_code: number;
    status:    string;
    token:     string;
    ci:        string;
    city:      string;
    password:  string;
    state:     string;
}